+++ packages/theme_tailor/example/lib/build_yaml_themes.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
-
-part 'build_yaml_themes.tailor.dart';
-
-@tailor
-class $_AnotherTheme {
-  static List<Color> themes = [Colors.amber, Colors.blueGrey.shade800];
-}