+++ packages/theme_tailor/example/test/diagnosticable_test.dart
@@ -1,8 +1,7 @@
 import 'dart:math';
 
 import 'package:example/diagnosticable.dart' as diagnosticable;
-import 'package:example/diagnosticable_barrel_import.dart'
-    as diagnosticable_barrel;
+import 'package:example/diagnosticable_barrel_import.dart' as diagnosticable_barrel;
 import 'package:example/diagnosticable_lib.dart';
 import 'package:example/empty_theme.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -22,20 +21,15 @@ void main() {
     expect(theme, isA<DiagnosticableTree>());
 
     expect(properties.properties, [
-      isA<DiagnosticsProperty>()
-          .having((d) => d.name, 'name', 'type')
-          .having((d) => d.value, 'value', 'MyTheme'),
+      isA<DiagnosticsProperty>().having((d) => d.name, 'name', 'type').having((d) => d.value, 'value', 'MyTheme'),
       isA<DiagnosticsProperty>()
           .having((d) => d.name, 'name', 'background')
           .having((d) => d.value, 'value', randomColor),
-      isA<DiagnosticsProperty>()
-          .having((d) => d.name, 'name', 'textStyle')
-          .having((d) => d.value, 'value', textStyle),
+      isA<DiagnosticsProperty>().having((d) => d.name, 'name', 'textStyle').having((d) => d.value, 'value', textStyle),
     ]);
   });
 
-  test('has debugFillProperties if importing flutter#foundation from barrel',
-      () {
+  test('has debugFillProperties if importing flutter#foundation from barrel', () {
     final properties = DiagnosticPropertiesBuilder();
 
     final randomColor = Colors.accents[Random().nextInt(Colors.accents.length)];
@@ -49,20 +43,16 @@ void main() {
     expect(theme, isA<DiagnosticableTree>());
 
     expect(properties.properties, [
-      isA<DiagnosticsProperty>()
-          .having((d) => d.name, 'name', 'type')
-          .having((d) => d.value, 'value', 'MyTheme'),
+      isA<DiagnosticsProperty>().having((d) => d.name, 'name', 'type').having((d) => d.value, 'value', 'MyTheme'),
       isA<DiagnosticsProperty>()
           .having((d) => d.name, 'name', 'background')
           .having((d) => d.value, 'value', randomColor),
-      isA<DiagnosticsProperty>()
-          .having((d) => d.name, 'name', 'textStyle')
-          .having((d) => d.value, 'value', textStyle),
+      isA<DiagnosticsProperty>().having((d) => d.name, 'name', 'textStyle').having((d) => d.value, 'value', textStyle),
     ]);
   });
 
   test('no debugFillProperties if not importing flutter#foundation', () {
-    final theme = EmptyTheme();
+    final theme = const EmptyTheme();
 
     expect(theme, isNot(isA<DiagnosticableTree>()));
   });