+++ packages/theme_tailor/lib/src/generator/theme_tailor_generator.dart
@@ -20,7 +20,6 @@ import 'package:theme_tailor/src/util/extension/element_annotation_extension.dar
 import 'package:theme_tailor/src/util/extension/element_extension.dart';
 import 'package:theme_tailor/src/util/extension/field_declaration_extension.dart';
 import 'package:theme_tailor/src/util/extension/library_element_extension.dart';
-import 'package:theme_tailor/src/util/extension/scope_extension.dart';
 import 'package:theme_tailor/src/util/field_helper.dart';
 import 'package:theme_tailor/src/util/string_format.dart';
 import 'package:theme_tailor/src/util/theme_encoder_helper.dart';
@@ -65,7 +64,16 @@ class ThemeTailorGenerator extends GeneratorForAnnotation<Tailor> {
     final fieldsToCheck =
         fields.values.where((f) => f.isTailorThemeExtension).map((f) => f.name);
 
-    final astVisitor = _TailorClassASTVisitor(fieldNamesToCheck: fieldsToCheck);
+    final typeDefAstVisitor = _TypeDefAstVisitor();
+    for (final unit in _getLibrariesCompilationUnits(
+        [library, ...library.importedLibraries])) {
+      unit.visitChildren(typeDefAstVisitor);
+    }
+
+    final astVisitor = _TailorClassASTVisitor(
+      fieldNamesToCheck: fieldsToCheck.toList(),
+      typeDefinitions: typeDefAstVisitor.typeDefinitions,
+    );
     _getAstNodeFromElement(element).visitChildren(astVisitor);
 
     for (final typeEntry in astVisitor.fieldTypes.entries) {
@@ -78,10 +86,15 @@ class ThemeTailorGenerator extends GeneratorForAnnotation<Tailor> {
     for (var i = 0; i < element.metadata.length; i++) {
       final annotation = element.metadata[i];
 
-      extractThemeEncoderData(
+      final encoder = extractThemeEncoderData(
         annotation,
         annotation.computeConstantValue()!,
-      )?.let((it) => classLevelEncoders[it.type] = it);
+      );
+
+      if (encoder != null) {
+        classLevelEncoders[encoder.type] = encoder;
+        continue;
+      }
 
       if (!annotation.isTailorAnnotation) {
         classLevelAnnotations.add(astVisitor.rawClassAnnotations[i]);
@@ -179,8 +192,13 @@ class _TailorClassVisitor extends SimpleElementVisitor {
   final extensionAnnotationTypeChecker =
       TypeChecker.fromRuntime(themeExtension.runtimeType);
 
+  final ignoreAnnotationTypeChecker =
+      TypeChecker.fromRuntime(ignore.runtimeType);
+
   @override
   void visitFieldElement(FieldElement element) {
+    if (ignoreAnnotationTypeChecker.hasAnnotationOf(element)) return;
+
     if (element.isStatic && element.type.isDartCoreList) {
       final propName = element.name;
       final isInternalAnnotation = <bool>[];
@@ -225,12 +243,16 @@ class _TailorClassVisitor extends SimpleElementVisitor {
 }
 
 class _TailorClassASTVisitor extends SimpleAstVisitor {
-  _TailorClassASTVisitor({required this.fieldNamesToCheck});
+  _TailorClassASTVisitor({
+    required this.fieldNamesToCheck,
+    required this.typeDefinitions,
+  });
 
   final List<String> rawClassAnnotations = [];
   final Map<String, List<String>> rawFieldsAnnotations = {};
 
-  final Iterable<String> fieldNamesToCheck;
+  final List<String> fieldNamesToCheck;
+  final Map<String, TypeAnnotation> typeDefinitions;
   final Map<String, String> fieldTypes = {};
 
   @override
@@ -246,8 +268,14 @@ class _TailorClassASTVisitor extends SimpleAstVisitor {
     rawFieldsAnnotations[fieldName] = node.annotations;
 
     if (fieldType != null && fieldNamesToCheck.contains(fieldName)) {
+      final typeDefinitionChildEntities =
+          typeDefinitions[node.fields.childEntities.first.toString()]
+              ?.childEntities;
+
       final childTypeEntities =
-          fieldType.childEntities.map((e) => e.toString()).toList();
+          (typeDefinitionChildEntities ?? fieldType.childEntities)
+              .map((e) => e.toString())
+              .toList();
       if (childTypeEntities.length >= 2 && childTypeEntities[0] == 'List') {
         final typeWithBraces = childTypeEntities[1];
         fieldTypes[fieldName] =
@@ -258,8 +286,37 @@ class _TailorClassASTVisitor extends SimpleAstVisitor {
 }
 
 AstNode _getAstNodeFromElement(Element element) {
-  final library = element.library!;
-  final result = library.session.getParsedLibraryByElement(library)
-      as ParsedLibraryResult?;
+  final result = _getParsedLibraryResultFromElement(element);
   return result!.getElementDeclaration(element)!.node;
+}
+
+List<CompilationUnit> _getLibrariesCompilationUnits(
+    List<LibraryElement> libraries) {
+  return libraries
+      .map(_getParsedLibraryResultFromElement)
+      .whereNotNull()
+      .map((lib) => lib.units.map((u) => u.unit))
+      .flattened
+      .toList();
+}
+
+ParsedLibraryResult? _getParsedLibraryResultFromElement(Element element) {
+  final library = element.library!;
+  final parsedLibrary = library.session.getParsedLibraryByElement(library);
+  if (parsedLibrary is ParsedLibraryResult) {
+    return parsedLibrary;
+  } else {
+    return null;
+  }
+}
+
+class _TypeDefAstVisitor extends SimpleAstVisitor {
+  final typeDefinitions = <String, TypeAnnotation>{};
+
+  @override
+  void visitGenericTypeAlias(GenericTypeAlias node) {
+    typeDefinitions[node.name.toString().replaceAll('?', '')] = node.type;
+
+    return super.visitGenericTypeAlias(node);
+  }
 }