+++ packages/theme_tailor_annotation/CHANGELOG.md
@@ -1,8 +1,11 @@
+# 1.0.3
+- Build configuration. It is possible to configure `themes` within the build.yaml
+
 # 1.0.2
-* Fixed encoder declaration that prevented creating encoders for nullable types
+- Fixed encoder declaration that prevented creating encoders for nullable types
 
 # 1.0.1
-* Exports `DeepCollectionEquality` from `collection` package to support `hashCode` and `== operator` overrides
+- Exports `DeepCollectionEquality` from `collection` package to support `hashCode` and `== operator` overrides
 
 # 1.0.0
-* Initial version of theme_tailor_annotation library
+- Initial version of theme_tailor_annotation library